Initial Refactoring Save User

+++ cliente/Cliente.Api/Domain/User.cs
@@ -2,7 +2,7 @@ namespace cliente.Cliente.Api.Domain
 {
     public class User
     {
-        public int Age { get; set; }
+        public short Age { get; set; }
         public string Name { get; set; }
     }
 }
